+++ tests/Unit/Bus/Results/ListOfErrorsTest.php
@@ -20,7 +20,7 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Tests\Unit\Bus\Results;
 
 use CloudCreativity\Modules\Bus\Results\Error;
-use CloudCreativity\Modules\Bus\Results\ErrorIterableInterface;
+use CloudCreativity\Modules\Bus\Results\ListOfErrorsInterface;
 use CloudCreativity\Modules\Bus\Results\KeyedSetOfErrors;
 use CloudCreativity\Modules\Bus\Results\ListOfErrors;
 use CloudCreativity\Modules\Tests\Unit\Infrastructure\Log\TestEnum;
@@ -38,10 +38,9 @@ class ListOfErrorsTest extends TestCase
             $b = new Error(null, 'Message B'),
         );
 
-        $this->assertInstanceOf(ErrorIterableInterface::class, $errors);
+        $this->assertInstanceOf(ListOfErrorsInterface::class, $errors);
         $this->assertSame([$a, $b], iterator_to_array($errors));
         $this->assertSame([$a, $b], $errors->all());
-        $this->assertSame($errors, $errors->toList());
         $this->assertEquals(new KeyedSetOfErrors($a, $b), $errors->toKeyedSet());
         $this->assertCount(2, $errors);
         $this->assertTrue($errors->isNotEmpty());